Rename middleware file
